quiz_agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiolimiter==1.2.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
base58==2.1.1
beautifulsoup4==4.13.3
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
cryptography==43.0.3
cytoolz==1.0.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
dydantic==0.0.8
email_validator==2.2.0
eth-hash==0.7.1
eth-keys==0.6.1
eth-typing==5.2.1
eth-utils==5.3.0
executing==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx
httpx-sse==0.4.0
idna==3.10
isoduration==20.11.0
jiter==0.9.0
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.29.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.53
langchain-google-genai==2.1.3
langchain-openai==0.3.8
langchain-text-splitters==0.3.7
loguru==0.7.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistune==3.1.3
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.3
oauthlib==3.2.2
openai
orjson==3.10.15
ormsgpack==1.9.0
overrides==7.7.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
redis==5.0.7 # Added for Redis caching
pure_eval==0.2.3
py-near
py_near_primitives==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-telegram-bot
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.24.0
rsa==4.9
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.1
structlog==25.2.0
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trustcall==0.0.38
tweepy==4.15.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.3.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
yarl==1.18.3
zstandard==0.23.0
